2004-10-05 Dario Massarin <nekkar@libero.it>         * transferlist.{h,cpp}: Api semplifications. In this moment there are some         new problems I have just discover and I will fix tomorrow.
        * transferlist.{h,cpp}, transfer.{h,cpp}: The way kget saves the transfers
        into the transfers.kgt file has been completely rethinked. The code has been
        rewritten from scratch and now it uses QDom. In this way each new transfer can,
        if needed, save its own data that isn't already saved by the Transfer class.

        * some fixes..

svn path=/branches/make_kget_cool/kdenetwork/kget/; revision=352195

+++ transferKio.cpp
@@ -12,6 +12,14 @@ TransferKio::TransferKio(Scheduler * _scheduler, const KURL & _src, const KURL &
 
 }
 
+TransferKio::TransferKio(Scheduler * _scheduler, QDomDocument * doc, QDomNode * n)
+    : Transfer(_scheduler, doc, n),
+      copyjob(0)
+{
+
+
+}
+
 bool TransferKio::slotResume()
 {
     kdDebug() << "TransferKio::slotResume (1)" << endl;
@@ -35,8 +43,11 @@ bool TransferKio::slotResume()
 
 void TransferKio::slotStop()
 {
-    copyjob->kill(true);
-    copyjob=0;
+    if(copyjob)
+    {
+        copyjob->kill(true);
+        copyjob=0;
+    }
     
     info.status = St_Stopped;
     info.speed = 0;
